requirements.txt
tensorflow
flask
pillow
gunicorn